# src/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Visit Manager for Contact Tracing",
    "version": "1.0.0",
    "title": "Visit Manager"
  },
...
  "paths": {
...
    "/establishments/PDF/{establishmentId}": {
      "get": {
        "tags": [
          "establishments"
        ],
        "summary": "Get establishment PDF",
        "description": "Get establishment PDF",
        "operationId": "getEstablishmentPDF",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathEstablishmentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Establishment PDF generated",
            "content": {
              "application/pdf": {}
            }
          },
          "404": {
            "description": "Establishment not found"
          }
        }
      }
    },
    "/establishments/PDF/{establishmentId}/space/{spaceId}": {
      "get": {
        "tags": [
          "establishments"
        ],
        "summary": "Get single space PDF",
        "description": "Get single space PDF",
        "operationId": "getSingleSpacePDF",
        "parameters": [
          {
            "$ref": "#/components/parameters/pathEstablishmentId"
          },
          {
            "$ref": "#/components/parameters/pathSpaceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Space PDF generated",
            "content": {
              "application/pdf": {}
            }
          },
          "404": {
            "description": "Space not found"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
        "pathEstablishmentId": {
          "in": "path",
          "name": "establishmentId",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        },
        "pathSpaceId": {
          "in": "path",
          "name": "spaceId",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        },
...
    },
...
  }
}